+++ tests/integ/snowflake/ml/registry/model/registry_catboost_model_test.py
@@ -7,6 +7,7 @@ from absl.testing import absltest, parameterized
 from sklearn import datasets, model_selection
 
 from snowflake.ml.model import model_signature
+from snowflake.ml.model._model_composer.model_manifest import model_manifest_schema
 from tests.integ.snowflake.ml.registry.model import registry_model_test_base
 from tests.integ.snowflake.ml.test_utils import dataframe_utils
 
@@ -81,6 +82,11 @@ class TestRegistryCatBoostModelInteg(registry_model_test_base.RegistryModelTestB
                     cal_X_test,
                     lambda res: np.testing.assert_allclose(res.values, expected_explanations),
                 ),
+            },
+            function_type_assert={
+                "explain": model_manifest_schema.ModelMethodFunctionTypes.TABLE_FUNCTION,
+                "predict": model_manifest_schema.ModelMethodFunctionTypes.FUNCTION,
+                "predict_proba": model_manifest_schema.ModelMethodFunctionTypes.FUNCTION,
             },
         )
 
@@ -193,6 +199,11 @@ class TestRegistryCatBoostModelInteg(registry_model_test_base.RegistryModelTestB
                     lambda res: dataframe_utils.check_sp_df_res(res, explanation_df_expected, check_dtype=False),
                 ),
             },
+            function_type_assert={
+                "explain": model_manifest_schema.ModelMethodFunctionTypes.TABLE_FUNCTION,
+                "predict": model_manifest_schema.ModelMethodFunctionTypes.FUNCTION,
+                "predict_proba": model_manifest_schema.ModelMethodFunctionTypes.FUNCTION,
+            },
         )
 
     @parameterized.product(  # type: ignore[misc]
@@ -246,6 +257,12 @@ class TestRegistryCatBoostModelInteg(registry_model_test_base.RegistryModelTestB
             },
             options={"enable_explainability": True},
             signatures=sig,
+            function_type_assert={
+                "explain": model_manifest_schema.ModelMethodFunctionTypes.TABLE_FUNCTION,
+                "predict": model_manifest_schema.ModelMethodFunctionTypes.FUNCTION,
+                "predict_proba": model_manifest_schema.ModelMethodFunctionTypes.FUNCTION,
+                "predict_log_proba": model_manifest_schema.ModelMethodFunctionTypes.FUNCTION,
+            },
         )
 
         with self.assertRaisesRegex(
@@ -276,6 +293,11 @@ class TestRegistryCatBoostModelInteg(registry_model_test_base.RegistryModelTestB
                 },
                 signatures=sig,
                 additional_version_suffix="v2",
+                function_type_assert={
+                    "explain": model_manifest_schema.ModelMethodFunctionTypes.TABLE_FUNCTION,
+                    "predict": model_manifest_schema.ModelMethodFunctionTypes.FUNCTION,
+                    "predict_proba": model_manifest_schema.ModelMethodFunctionTypes.FUNCTION,
+                },
             )