+++ tests/integ/snowflake/ml/registry/services/registry_model_deployment_test_base.py
@@ -1,18 +1,23 @@
 import inspect
+import logging
 import os
 import pathlib
+import time
 import uuid
-from typing import Any, Callable, Dict, List, Optional, Tuple
+from typing import Any, Callable, Dict, List, Optional, Tuple, cast
 
 import pytest
 import yaml
 from absl.testing import absltest
+from packaging import version
 
 from snowflake.ml._internal import file_utils
 from snowflake.ml._internal.utils import snowflake_env, sql_identifier
 from snowflake.ml.model import ModelVersion, type_hints as model_types
+from snowflake.ml.model._client.ops import service_ops
 from snowflake.ml.model._client.service import model_deployment_spec
 from snowflake.ml.registry import registry
+from snowflake.snowpark import row
 from snowflake.snowpark._internal import utils as snowpark_utils
 from tests.integ.snowflake.ml.test_utils import (
     common_test_base,
@@ -126,9 +131,38 @@ class RegistryModelDeploymentTestBase(common_test_base.CommonTestBase):
         )
 
         # deploy the model service
-        mv._service_ops._service_client.deploy_model(
+        query_id, async_job = mv._service_ops._service_client.deploy_model(
             stage_path=stage_path, model_deployment_spec_file_rel_path=deploy_spec_file_rel_path
         )
+
+        # TODO(hayu): Remove the version check after Snowflake 8.37.0 release
+        if snowflake_env.get_current_snowflake_version(self.session) >= version.parse("8.37.0"):
+            # stream service logs in a thread
+            model_build_service_name = sql_identifier.SqlIdentifier(
+                mv._service_ops._get_model_build_service_name(query_id)
+            )
+            model_build_service = service_ops.ServiceLogInfo(
+                database_name=database_name,
+                schema_name=schema_name,
+                service_name=model_build_service_name,
+                container_name="model-build",
+            )
+            model_inference_service = service_ops.ServiceLogInfo(
+                database_name=database_name,
+                schema_name=schema_name,
+                service_name=sql_identifier.SqlIdentifier(service_name),
+                container_name="model-inference",
+            )
+            services = [model_build_service, model_inference_service]
+            log_thread = mv._service_ops._start_service_log_streaming(async_job, services, False, True)
+            log_thread.join()
+        else:
+            while not async_job.is_done():
+                time.sleep(5)
+
+        res = cast(str, cast(List[row.Row], async_job.result())[0][0])
+        logging.info(f"Inference service {service_name} deployment complete: {res}")
+        return res
 
     def _test_registry_model_deployment(
         self,